Change BlockEnvConstructor so it's a subtrait of ChainSpec
- This prevent having another type in addition to ChainSpec
- Also allows moving the trait back again to edr_evm crate

- Implemented trait definition for OpChainSpec
