+++ crates/edr_evm/src/spec.rs
@@ -1,12 +1,13 @@
 use std::{fmt::Debug, marker::PhantomData, sync::Arc};
 
 use edr_eth::{
-    block,
-    l1::{self, L1ChainSpec},
+    block::{self, BlobGas, Header, PartialHeader},
+    eips::eip4844,
+    l1::{self, BlockEnv, L1ChainSpec},
     log::{ExecutionLog, FilterLog},
     receipt::{BlockReceipt, ExecutionReceipt, MapReceiptLogs, ReceiptTrait},
     result::ResultAndState,
-    spec::{BlockEnvConstructor, ChainHardfork, ChainSpec, EthHeaderConstants},
+    spec::{ChainHardfork, ChainSpec, EthHeaderConstants},
     Bytes, B256,
 };
 use edr_rpc_eth::{spec::RpcSpec, RpcTypeFrom, TransactionConversionError};
@@ -162,8 +163,8 @@ pub trait RuntimeSpec:
           + ExecutableTransaction
           + TransactionType
           + TransactionValidation<ValidationError: From<l1::InvalidTransaction>>,
-        BlockConstructor: BlockEnvConstructor<block::PartialHeader, Self::BlockEnv> + BlockEnvConstructor<block::Header, Self::BlockEnv>,
     >
+    + BlockEnvConstructor<block::PartialHeader> + BlockEnvConstructor<block::Header>
     // Defines an RPC spec and conversion between RPC <-> EVM types
     + RpcSpec<
         RpcBlock<<Self as RpcSpec>::RpcTransaction>: EthRpcBlock
@@ -327,6 +328,12 @@ pub trait RuntimeSpec:
 
 }
 
+/// A trait for constructing a (partial) block header into an EVM block.
+pub trait BlockEnvConstructor<HeaderT>: ChainSpec {
+    /// Converts the instance into an EVM block.
+    fn new_block_env(header: &HeaderT, hardfork: l1::SpecId) -> Self::BlockEnv;
+}
+
 /// A supertrait for [`RuntimeSpec`] that is safe to send between threads.
 pub trait SyncRuntimeSpec:
     RuntimeSpec<
@@ -453,5 +460,132 @@ impl RuntimeSpec for L1ChainSpec {
             instruction: EthInstructions::default(),
             precompiles: precompile_provider,
         }
+    }
+}
+
+impl BlockEnvConstructor<PartialHeader> for L1ChainSpec {
+    fn new_block_env(header: &PartialHeader, hardfork: l1::SpecId) -> Self::BlockEnv {
+        BlockEnv {
+            number: header.number,
+            beneficiary: header.beneficiary,
+            timestamp: header.timestamp,
+            difficulty: header.difficulty,
+            basefee: header.base_fee.map_or(0u64, |base_fee| {
+                base_fee.try_into().expect("base fee is too large")
+            }),
+            gas_limit: header.gas_limit,
+            prevrandao: if hardfork >= l1::SpecId::MERGE {
+                Some(header.mix_hash)
+            } else {
+                None
+            },
+            blob_excess_gas_and_price: header.blob_gas.as_ref().map(
+                |BlobGas { excess_gas, .. }| {
+                    eip4844::BlobExcessGasAndPrice::new(*excess_gas, hardfork >= l1::SpecId::PRAGUE)
+                },
+            ),
+        }
+    }
+}
+
+impl BlockEnvConstructor<Header> for L1ChainSpec {
+    fn new_block_env(header: &Header, hardfork: l1::SpecId) -> Self::BlockEnv {
+        BlockEnv {
+            number: header.number,
+            beneficiary: header.beneficiary,
+            timestamp: header.timestamp,
+            difficulty: header.difficulty,
+            basefee: header.base_fee_per_gas.map_or(0u64, |base_fee| {
+                base_fee.try_into().expect("base fee is too large")
+            }),
+            gas_limit: header.gas_limit,
+            prevrandao: if hardfork >= l1::SpecId::MERGE {
+                Some(header.mix_hash)
+            } else {
+                None
+            },
+            blob_excess_gas_and_price: header.blob_gas.as_ref().map(
+                |BlobGas { excess_gas, .. }| {
+                    eip4844::BlobExcessGasAndPrice::new(*excess_gas, hardfork >= l1::SpecId::PRAGUE)
+                },
+            ),
+        }
+    }
+}
+
+#[cfg(test)]
+mod l1_chain_spec_tests {
+
+    use edr_eth::{
+        block::{BlobGas, Header},
+        l1, Address, Bloom, Bytes, B256, B64, U256,
+    };
+
+    use crate::spec::{BlockEnvConstructor as _, L1ChainSpec};
+
+    fn build_block_header(blob_gas: Option<BlobGas>) -> Header {
+        Header {
+            parent_hash: B256::default(),
+            ommers_hash: B256::default(),
+            beneficiary: Address::default(),
+            state_root: B256::default(),
+            transactions_root: B256::default(),
+            receipts_root: B256::default(),
+            logs_bloom: Bloom::default(),
+            difficulty: U256::default(),
+            number: 124,
+            gas_limit: u64::default(),
+            gas_used: 1337,
+            timestamp: 0,
+            extra_data: Bytes::default(),
+            mix_hash: B256::default(),
+            nonce: B64::from(99u64),
+            base_fee_per_gas: None,
+            withdrawals_root: None,
+            blob_gas,
+            parent_beacon_block_root: None,
+            requests_hash: Some(B256::random()),
+        }
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_for_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = L1ChainSpec::new_block_env(&header, l1::SpecId::CANCUN);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_below_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = L1ChainSpec::new_block_env(&header, l1::SpecId::SHANGHAI);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_above_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = L1ChainSpec::new_block_env(&header, l1::SpecId::PRAGUE);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_use_existing_excess_blob_gas() {
+        let excess_gas = 0x80000u64;
+        let blob_gas = BlobGas {
+            excess_gas,
+            gas_used: 0x80000u64,
+        };
+        let header = build_block_header(Some(blob_gas)); // blob gas present
+
+        let block = L1ChainSpec::new_block_env(&header, l1::SpecId::CANCUN);
+
+        let blob_excess_gas = block
+            .blob_excess_gas_and_price
+            .expect("Blob excess gas should be set");
+        assert_eq!(blob_excess_gas.excess_blob_gas, excess_gas);
     }
 }
