+++ crates/edr_op/src/spec.rs
@@ -3,15 +3,19 @@ use std::sync::Arc;
 
 use alloy_rlp::RlpEncodable;
 use edr_eth::{
-    eips::eip1559::{BaseFeeParams, ConstantBaseFeeParams, ForkBaseFeeParams},
-    l1,
+    block::{BlobGas, Header, PartialHeader},
+    eips::{
+        eip1559::{BaseFeeParams, ConstantBaseFeeParams, ForkBaseFeeParams},
+        eip4844,
+    },
+    l1::{self, BlockEnv},
     spec::{ChainHardfork, ChainSpec, EthHeaderConstants},
 };
 use edr_evm::{
     evm::Evm,
     interpreter::{EthInstructions, EthInterpreter, InterpreterResult},
     precompile::PrecompileProvider,
-    spec::{ContextForChainSpec, GenesisBlockFactory, RuntimeSpec},
+    spec::{BlockEnvConstructor, ContextForChainSpec, GenesisBlockFactory, RuntimeSpec},
     state::Database,
     transaction::{TransactionError, TransactionErrorForChainSpec, TransactionValidation},
     BlockReceipts, EthLocalBlockForChainSpec, LocalCreationError, RemoteBlock,
@@ -63,7 +67,6 @@ impl ChainSpec for OpChainSpec {
     type Context = L1BlockInfo;
     type HaltReason = OpHaltReason;
     type SignedTransaction = transaction::Signed;
-    type BlockConstructor = l1::L1BlockConstructor;
 }
 
 impl GenesisBlockFactory for OpChainSpec {
@@ -241,5 +244,133 @@ impl<TimerT: Clone + TimeSinceEpoch> ProviderSpec<TimerT> for OpChainSpec {
             },
             remainder @ OpHaltReason::FailedDeposit => TransactionFailureReason::Inner(remainder),
         }
+    }
+}
+
+impl BlockEnvConstructor<PartialHeader> for OpChainSpec {
+    fn new_block_env(header: &PartialHeader, hardfork: l1::SpecId) -> Self::BlockEnv {
+        BlockEnv {
+            number: header.number,
+            beneficiary: header.beneficiary,
+            timestamp: header.timestamp,
+            difficulty: header.difficulty,
+            basefee: header.base_fee.map_or(0u64, |base_fee| {
+                base_fee.try_into().expect("base fee is too large")
+            }),
+            gas_limit: header.gas_limit,
+            prevrandao: if hardfork >= l1::SpecId::MERGE {
+                Some(header.mix_hash)
+            } else {
+                None
+            },
+            blob_excess_gas_and_price: header.blob_gas.as_ref().map(
+                |BlobGas { excess_gas, .. }| {
+                    eip4844::BlobExcessGasAndPrice::new(*excess_gas, hardfork >= l1::SpecId::PRAGUE)
+                },
+            ),
+        }
+    }
+}
+
+impl BlockEnvConstructor<Header> for OpChainSpec {
+    fn new_block_env(header: &Header, hardfork: l1::SpecId) -> Self::BlockEnv {
+        BlockEnv {
+            number: header.number,
+            beneficiary: header.beneficiary,
+            timestamp: header.timestamp,
+            difficulty: header.difficulty,
+            basefee: header.base_fee_per_gas.map_or(0u64, |base_fee| {
+                base_fee.try_into().expect("base fee is too large")
+            }),
+            gas_limit: header.gas_limit,
+            prevrandao: if hardfork >= l1::SpecId::MERGE {
+                Some(header.mix_hash)
+            } else {
+                None
+            },
+            blob_excess_gas_and_price: header.blob_gas.as_ref().map(
+                |BlobGas { excess_gas, .. }| {
+                    eip4844::BlobExcessGasAndPrice::new(*excess_gas, hardfork >= l1::SpecId::PRAGUE)
+                },
+            ),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    use edr_eth::{
+        block::{BlobGas, Header},
+        l1, Address, Bloom, Bytes, B256, B64, U256,
+    };
+    use edr_evm::spec::BlockEnvConstructor as _;
+
+    use crate::spec::OpChainSpec;
+
+    fn build_block_header(blob_gas: Option<BlobGas>) -> Header {
+        Header {
+            parent_hash: B256::default(),
+            ommers_hash: B256::default(),
+            beneficiary: Address::default(),
+            state_root: B256::default(),
+            transactions_root: B256::default(),
+            receipts_root: B256::default(),
+            logs_bloom: Bloom::default(),
+            difficulty: U256::default(),
+            number: 124,
+            gas_limit: u64::default(),
+            gas_used: 1337,
+            timestamp: 0,
+            extra_data: Bytes::default(),
+            mix_hash: B256::default(),
+            nonce: B64::from(99u64),
+            base_fee_per_gas: None,
+            withdrawals_root: None,
+            blob_gas,
+            parent_beacon_block_root: None,
+            requests_hash: Some(B256::random()),
+        }
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_for_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = OpChainSpec::new_block_env(&header, l1::SpecId::CANCUN);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_below_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = OpChainSpec::new_block_env(&header, l1::SpecId::SHANGHAI);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_not_default_excess_blob_gas_above_cancun() {
+        let header = build_block_header(None); // No blob gas information
+
+        let block = OpChainSpec::new_block_env(&header, l1::SpecId::PRAGUE);
+        assert_eq!(block.blob_excess_gas_and_price, None);
+    }
+
+    #[test]
+    fn generic_block_constructor_should_use_existing_excess_blob_gas() {
+        let excess_gas = 0x80000u64;
+        let blob_gas = BlobGas {
+            excess_gas,
+            gas_used: 0x80000u64,
+        };
+        let header = build_block_header(Some(blob_gas)); // blob gas present
+
+        let block = OpChainSpec::new_block_env(&header, l1::SpecId::CANCUN);
+
+        let blob_excess_gas = block
+            .blob_excess_gas_and_price
+            .expect("Blob excess gas should be set");
+        assert_eq!(blob_excess_gas.excess_blob_gas, excess_gas);
     }
 }
