+++ crates/edr_provider/src/data/call.rs
@@ -1,6 +1,6 @@
 use edr_eth::{
-    block::Header, l1, result::ExecutionResult, spec::BlockEnvConstructor,
-    transaction::TransactionValidation, Address, HashMap,
+    block::Header, l1, result::ExecutionResult, transaction::TransactionValidation, Address,
+    HashMap,
 };
 use edr_evm::{
     blockchain::{BlockHash, BlockchainErrorForChainSpec},
@@ -44,7 +44,7 @@ where
     let mut header = header.clone();
     header.base_fee_per_gas = header.base_fee_per_gas.map(|_| 0);
 
-    let block = ChainSpecT::BlockConstructor::new_block_env(&header, cfg_env.spec.into());
+    let block = ChainSpecT::new_block_env(&header, cfg_env.spec.into());
 
     guaranteed_dry_run_with_inspector::<_, ChainSpecT, _, _>(
         blockchain,
